backend/requirements.txt
bcrypt==4.1.2
bidict==0.23.1
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SocketIO==5.3.5
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.4.0
pandas==2.3.3
psycopg2==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-engineio==4.13.0
python-socketio==5.10.0
pytz==2025.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
tzdata==2025.3
Werkzeug==3.1.4
wheel==0.45.1
wsproto==1.3.2